Buttons in the Top - Right corner
I think it's more convenient if the buttons are in upper Top-Right corner like in windows, because you have "infinite reachability". And if you could manage to show the button on the second display ( when you have the Multi monitors add-on")  I will be very thankful. Then you have created the perfect ubuntu 18 experience.